refactor: split controller.go into cohesive domain files
Extract ~1,770 lines from the monolithic controller.go (2,489 → 725 lines)
into focused files organized by domain:

- shutdown.go: lifecycle management, signal handling, VM termination
- logging.go: log methods, token logging, tracer init
- init.go: workspace setup, credentials, cloning, prompt loading
- issue_types.go: shared issue/label/comment types
- monorepo.go: package scope validation and instructions
- issues.go: issue fetching, existing work detection, branch naming
- prompt.go: prompt building, rendering, feedback sections
- task_state.go: phase transitions, termination checks, final logs
- iteration.go: runIteration, fallback params
- dependencies.go: merged controller-level dependency methods

Tests migrated to companion *_test.go files alongside their source.
No behavioral changes - pure file-move refactoring within the controller package.

Closes #455
